Separate out freespec tests, and runs them under dotty build as well (skipped some tests).
