+++ build.sbt
@@ -18,6 +18,7 @@ lazy val scalatestTest = ScalatestBuild.scalatestTest
 lazy val scalatestDiagramsTest = ScalatestBuild.scalatestDiagramsTest
 lazy val scalatestFeatureSpecTest = ScalatestBuild.scalatestFeatureSpecTest
 lazy val scalatestFlatSpecTest = ScalatestBuild.scalatestFlatSpecTest
+lazy val scalatestFreeSpecTest = ScalatestBuild.scalatestFreeSpecTest
 lazy val scalatestJS = ScalatestBuild.scalatestJS
 lazy val scalatestNative = ScalatestBuild.scalatestNative
 lazy val scalatestDotty = ScalatestBuild.scalatestDotty
@@ -25,14 +26,17 @@ lazy val scalatestTestJS = ScalatestBuild.scalatestTestJS
 lazy val scalatestDiagramsTestJS = ScalatestBuild.scalatestDiagramsTestJS
 lazy val scalatestFeatureSpecTestJS = ScalatestBuild.scalatestFeatureSpecTestJS
 lazy val scalatestFlatSpecTestJS = ScalatestBuild.scalatestFlatSpecTestJS
+lazy val scalatestFreeSpecTestJS = ScalatestBuild.scalatestFreeSpecTestJS
 lazy val scalatestTestNative = ScalatestBuild.scalatestTestNative
 lazy val scalatestDiagramsTestNative = ScalatestBuild.scalatestDiagramsTestNative
 lazy val scalatestFeatureSpecTestNative = ScalatestBuild.scalatestFeatureSpecTestNative
 lazy val scalatestFlatSpecTestNative = ScalatestBuild.scalatestFlatSpecTestNative
+lazy val scalatestFreeSpecTestNative = ScalatestBuild.scalatestFreeSpecTestNative
 lazy val scalatestTestDotty = ScalatestBuild.scalatestTestDotty
 lazy val scalatestDiagramsTestDotty = ScalatestBuild.scalatestDiagramsTestDotty
 lazy val scalatestFeatureSpecTestDotty = ScalatestBuild.scalatestFeatureSpecTestDotty
 lazy val scalatestFlatSpecTestDotty = ScalatestBuild.scalatestFlatSpecTestDotty
+lazy val scalatestFreeSpecTestDotty = ScalatestBuild.scalatestFreeSpecTestDotty
 lazy val scalatestApp = ScalatestBuild.scalatestApp
 lazy val scalatestAppJS = ScalatestBuild.scalatestAppJS
 lazy val scalatestAppNative = ScalatestBuild.scalatestAppNative
