+++ jvm/freespec-test/src/test/scala/org/scalatest/freespec/FixtureAsyncFreeSpecLikeSpec2.scala
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.scalatest.fixture
+package org.scalatest.freespec
 
 import org.scalatest._
 import SharedHelpers.EventRecordingReporter
@@ -25,13 +25,15 @@ import scala.util.Success
 import org.scalatest
 import org.scalatest.freespec
 
-class AsyncFreeSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
+class FixtureAsyncFreeSpecLikeSpec2 extends scalatest.funspec.AsyncFunSpec {
 
-  describe("AsyncFreeSpec") {
+  describe("AsyncFreeSpecLike") {
 
+    // SKIP-DOTTY-START
+    // ParallelTestExecution not working yet.
     it("can be used for tests that return Future under parallel async test execution") {
 
-      class ExampleSpec extends freespec.FixtureAsyncFreeSpec with ParallelTestExecution {
+      class ExampleSpec extends freespec.FixtureAsyncFreeSpecLike with ParallelTestExecution {
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -59,13 +61,13 @@ class AsyncFreeSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
 
         "test 4" in { fixture =>
           Future {
-            cancel
+            cancel()
           }
         }
 
         "test 5" ignore { fixture =>
           Future {
-            cancel
+            cancel()
           }
         }
 
@@ -94,7 +96,7 @@ class AsyncFreeSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
 
     it("can be used for tests that did not return Future under parallel async test execution") {
 
-      class ExampleSpec extends freespec.FixtureAsyncFreeSpec with ParallelTestExecution {
+      class ExampleSpec extends freespec.FixtureAsyncFreeSpecLike with ParallelTestExecution {
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -115,11 +117,11 @@ class AsyncFreeSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
         }
 
         "test 4" in { fixture =>
-          cancel
+          cancel()
         }
 
         "test 5" ignore { fixture =>
-          cancel
+          cancel()
         }
 
         override def newInstance = new ExampleSpec
@@ -144,12 +146,13 @@ class AsyncFreeSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
         assert(repo.testIgnoredEventsReceived(0).testName == "test 5")
       }
     }
+    // SKIP-DOTTY-END
 
     it("should run tests that return Future in serial by default") {
 
       @volatile var count = 0
 
-      class ExampleSpec extends freespec.FixtureAsyncFreeSpec {
+      class ExampleSpec extends freespec.FixtureAsyncFreeSpecLike {
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -196,7 +199,7 @@ class AsyncFreeSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
 
       @volatile var count = 0
 
-      class ExampleSpec extends freespec.FixtureAsyncFreeSpec {
+      class ExampleSpec extends freespec.FixtureAsyncFreeSpecLike {
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -219,7 +222,7 @@ class AsyncFreeSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
         "test 3" in { fixture =>
           assert(count == 2)
         }
-
+        
       }
 
       val rep = new EventRecordingReporter
@@ -241,7 +244,7 @@ class AsyncFreeSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
       var test2Thread: Option[Thread] = None
       var onCompleteThread: Option[Thread] = None
 
-      class ExampleSpec extends freespec.FixtureAsyncFreeSpec {
+      class ExampleSpec extends freespec.FixtureAsyncFreeSpecLike {
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -288,7 +291,7 @@ class AsyncFreeSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
       @volatile var test2Thread: Option[Thread] = None
       var onCompleteThread: Option[Thread] = None
 
-      class ExampleSpec extends freespec.FixtureAsyncFreeSpec {
+      class ExampleSpec extends freespec.FixtureAsyncFreeSpecLike {
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -351,7 +354,7 @@ class AsyncFreeSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
 
     it("should not run out of stack space with nested futures when using SerialExecutionContext") {
 
-      class ExampleSpec extends freespec.FixtureAsyncFreeSpec {
+      class ExampleSpec extends freespec.FixtureAsyncFreeSpecLike {
 
         // Note we get a StackOverflowError with the following execution
         // context.
@@ -387,7 +390,7 @@ class AsyncFreeSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
 
     it("should run tests that returns Future and report their result in serial") {
 
-      class ExampleSpec extends freespec.FixtureAsyncFreeSpec {
+      class ExampleSpec extends freespec.FixtureAsyncFreeSpecLike {
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -435,7 +438,7 @@ class AsyncFreeSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
 
     it("should run tests that does not return Future and report their result in serial") {
 
-      class ExampleSpec extends freespec.FixtureAsyncFreeSpec {
+      class ExampleSpec extends freespec.FixtureAsyncFreeSpecLike {
 
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
@@ -476,7 +479,7 @@ class AsyncFreeSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send an InfoProvided event for an info in main spec body") {
-      class MySuite extends freespec.FixtureAsyncFreeSpec  {
+      class MySuite extends freespec.FixtureAsyncFreeSpecLike  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -499,7 +502,7 @@ class AsyncFreeSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send an InfoProvided event for an info in scope body") {
-      class MySuite extends freespec.FixtureAsyncFreeSpec  {
+      class MySuite extends freespec.FixtureAsyncFreeSpecLike  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -526,7 +529,7 @@ class AsyncFreeSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send an InfoProvided event for an info in test body") {
-      class MySuite extends freespec.FixtureAsyncFreeSpec  {
+      class MySuite extends freespec.FixtureAsyncFreeSpecLike  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -558,7 +561,7 @@ class AsyncFreeSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send an InfoProvided event for an info in Future returned by test body") {
-      class MySuite extends freespec.FixtureAsyncFreeSpec  {
+      class MySuite extends freespec.FixtureAsyncFreeSpecLike  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -592,7 +595,7 @@ class AsyncFreeSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send a NoteProvided event for a note in main spec body") {
-      class MySuite extends freespec.FixtureAsyncFreeSpec  {
+      class MySuite extends freespec.FixtureAsyncFreeSpecLike  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -615,7 +618,7 @@ class AsyncFreeSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send a NoteProvided event for a note in scope body") {
-      class MySuite extends freespec.FixtureAsyncFreeSpec  {
+      class MySuite extends freespec.FixtureAsyncFreeSpecLike  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -642,7 +645,7 @@ class AsyncFreeSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send a NoteProvided event for a note in test body") {
-      class MySuite extends freespec.FixtureAsyncFreeSpec  {
+      class MySuite extends freespec.FixtureAsyncFreeSpecLike  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -667,7 +670,7 @@ class AsyncFreeSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send a NoteProvided event for a note in Future returned by test body") {
-      class MySuite extends freespec.FixtureAsyncFreeSpec  {
+      class MySuite extends freespec.FixtureAsyncFreeSpecLike  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -694,7 +697,7 @@ class AsyncFreeSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send an AlertProvided event for an alert in main spec body") {
-      class MySuite extends freespec.FixtureAsyncFreeSpec  {
+      class MySuite extends freespec.FixtureAsyncFreeSpecLike  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -717,7 +720,7 @@ class AsyncFreeSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send an AlertProvided event for an alert in scope body") {
-      class MySuite extends freespec.FixtureAsyncFreeSpec  {
+      class MySuite extends freespec.FixtureAsyncFreeSpecLike  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -744,7 +747,7 @@ class AsyncFreeSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send an AlertProvided event for an alert in test body") {
-      class MySuite extends freespec.FixtureAsyncFreeSpec  {
+      class MySuite extends freespec.FixtureAsyncFreeSpecLike  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -769,7 +772,7 @@ class AsyncFreeSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send an AlertProvided event for an alert in Future returned by test body") {
-      class MySuite extends freespec.FixtureAsyncFreeSpec  {
+      class MySuite extends freespec.FixtureAsyncFreeSpecLike  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -796,7 +799,7 @@ class AsyncFreeSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send a MarkupProvided event for a markup in main spec body") {
-      class MySuite extends freespec.FixtureAsyncFreeSpec  {
+      class MySuite extends freespec.FixtureAsyncFreeSpecLike  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -819,7 +822,7 @@ class AsyncFreeSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send a MarkupProvided event for a markup in scope body") {
-      class MySuite extends freespec.FixtureAsyncFreeSpec  {
+      class MySuite extends freespec.FixtureAsyncFreeSpecLike  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -846,7 +849,7 @@ class AsyncFreeSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send a MarkupProvided event for a markup in test body") {
-      class MySuite extends freespec.FixtureAsyncFreeSpec  {
+      class MySuite extends freespec.FixtureAsyncFreeSpecLike  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -878,7 +881,7 @@ class AsyncFreeSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should send a MarkupProvided event for a markup in Future returned by test body") {
-      class MySuite extends freespec.FixtureAsyncFreeSpec  {
+      class MySuite extends freespec.FixtureAsyncFreeSpecLike  {
         type FixtureParam = String
         def withFixture(test: OneArgAsyncTest): FutureOutcome =
           test("testing")
@@ -912,7 +915,7 @@ class AsyncFreeSpecSpec2 extends scalatest.funspec.AsyncFunSpec {
     }
 
     it("should allow other execution context to be used") {
-      class TestSpec extends freespec.FixtureAsyncFreeSpec {
+      class TestSpec extends freespec.FixtureAsyncFreeSpecLike {
         // SKIP-SCALATESTJS,NATIVE-START
         override implicit val executionContext = scala.concurrent.ExecutionContext.Implicits.global
         // SKIP-SCALATESTJS,NATIVE-END
