+++ project/GenScalaTestDotty.scala
@@ -285,4 +285,7 @@ object GenScalaTestDotty {
         )
       )    
 
+    def genFreeSpecTest(targetDir: File, version: String, scalaVersion: String): Seq[File] = 
+      copyDir("jvm/freespec-test/src/test/scala/org/scalatest/freespec", "org/scalatest/freespec", targetDir, List.empty)  
+
 }
